Add support to redirect to https when behind an SSL terminator
Just set varnish::vcl::https_redirect to true.

Varnish doesn't do SSL, but if it has a SSL terminator in front of it,
usually, it recieves a X-Forwarded-Proto: HTTP header telling it if the
request was over http or https.

With this flag turned on, requests over http:// are redirected to https://