Allow specifying selector ordering
Currently it is not possible to specify an order for Varnish selectors. This means that if multiple selectors satisfy a request then it is not possible to specify which selector "wins". This pull request adds a `$order` parameter to `varnish::selector` to fix this. I have also added some basic RSpec test coverage.