Add `X-Backend` and `X-Director` headers
In Varnish versions greater than or equal to 4.0, it is not possible to compare `req.backend_hint` to a string (see https://www.varnish-cache.org/lists/pipermail/varnish-bugs/2014-May/006051.html). The recommended workaround is to store the name of the backend hint in a custom header and use that custom header for comparisons instead.