Add support to use a custom package name
In certain scenarios it might make sense to use a different package name
than the hardcoded 'varnish' package name (eg. when using varnish-plus)

This commit introduces a new parameter $package_name which defaults to
'varnish'. The parameter is used in the install.pp class to set the name
of the package to install. Since the default is set to 'varnish', this
change does not break any existing setups.